server/server/requirements.txt
fastapi
uvicorn

langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters

sentence-transformers


huggingface-hub

pinecone-client==3.0.3

numpy
pydantic
python-dotenv
requests
tqdm
